serveur_fastapi/requirements.txt
mariadb==1.0.7
fastapi~=0.85.0
pydantic~=1.10.2
uvicorn~=0.18.3
websockets~=10.4
paho-mqtt~=1.6.1